Adds functions for setting up modern bash version
The `bash` that ships with OSX doesn't support many nice modern Bash
features (e.g. associative arrays, auto_cd), because it's six years old
and an entire major version back (3.2 vs 4.2).

So, if running the system bash, print out a warning message.  Then, if
desired, the user can choose to use homebrew to install a better version
and set it up as the default shell.