Adds function to override vundle default directory
When using the dotfiles from a user account that doesn't own the
repository, .vim/bundle isn't writable.  So, allow for the possibility
of using a different vundle bundle directory:

~/alt-vim-bundle

If the alt dir exists, it will be used.  Otherwise the function tries to
determine whether the default can be used.