Prefix DRIVER environment variable
The DRIVER environment variable was conflicting with the teaspoon
library's DRIVER environment variable. Teaspoon uses this variable to
determine the proper javascript runtime to use when testing. By
prefixing our variable with 'M_' we ensure we won't conflict with this
gem.

'M_' was used instead of '4M_' because 4 was not a valid character for the
name.